refactor: move useNavigateToNewOrderCallback to its own file
